added conditional to prevent admin account being deleted

+++ pages/api/admin/users/[id].js
@@ -17,9 +17,12 @@ const handler = async (req, res) => {
 };
 
 const deleteHandler = async (req, res) => {
-  console.log(req.query);
   let id = req.query.id;
-  console.log(id);
+  let userToDelete = await User.findById({ _id: new ObjectId(id) });
+
+  if (userToDelete.isAdmin) {
+    return res.status(400).send({ message: 'Can not delete ADMIN' });
+  }
   await User.findByIdAndDelete({ _id: new ObjectId(id) });
   console.log('user has been deleted');
   return res.status(201).send({ message: 'I hope this worked' });
